backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
python-multipart
httpx
python-dotenv
mp-api
